Moved Input File Handling to a Separate File to clean up code.
