+++ bin/InputFileHandler.h
@@ -1,0 +1,9 @@
+#ifndef LISPCOMPILER_INPUTFILEHANDLER_H
+#define LISPCOMPILER_INPUTFILEHANDLER_H
+
+#include "iostream"
+
+std::string fileToString(const std::string &fileName);
+
+
+#endif //LISPCOMPILER_INPUTFILEHANDLER_H
